refactor: move resolve_table from catalog to records toolset
resolve_table is a generic table-resolution utility (queries sys_db_object),
not catalog-specific. Its callers span catalog, ui-actions, and ui-policies,
so its home in catalog/read.ts was an accident of where it was first needed.
The records toolset (generic table CRUD) is its natural home.

Pure relocation — tool logic is unchanged.
