+++ src/tools/records/table-crud.ts
@@ -1,7 +1,12 @@
 import { z } from 'zod';
 import type { ServiceNowClient } from '../../clients/servicenow.js';
 import type { SnReference } from '../../types/servicenow.js';
-import { handleError, textResult } from '../helpers.js';
+import {
+  handleError,
+  resolveDisplay,
+  resolveValue,
+  textResult,
+} from '../helpers.js';
 import type { ToolRegistrar } from '../registry.js';
 
 export function registerTableCrudTools(
@@ -226,6 +231,93 @@ export function registerTableCrudTools(
         return textResult(
           `Updated record in "${table}" (sys_id: ${sys_id}) using ${patch ? 'PATCH' : 'PUT'}.`,
         );
+      } catch (err) {
+        return handleError(err);
+      }
+    },
+  );
+
+  registry.registerTool(
+    'resolve_table',
+    {
+      access: 'read',
+      title: 'Resolve Table sys_id',
+      description: [
+        'Looks up a ServiceNow table and returns its internal name and label.',
+        '',
+        'WHEN TO USE: before create_catalog_variable when type=8 (Reference) or',
+        'type=21 (List Collector) and you are unsure of the exact internal table name.',
+        "The internal name (e.g. 'sys_user_group') is what you pass to reference_table",
+        'or list_table — NOT the sys_id.',
+        '',
+        'Pass the exact internal name for a precise match, or a partial string to',
+        'search by name or label across all tables.',
+        '',
+        'Returns: internal name, human-readable label, and sys_id for each match.',
+      ].join('\n'),
+      inputSchema: {
+        name: z
+          .string()
+          .min(1)
+          .describe(
+            "Internal table name (e.g. 'sys_user') for an exact match, " +
+              'or a partial string to search by name or label.',
+          ),
+      },
+      annotations: {
+        openWorldHint: true,
+      },
+    },
+    async ({ name }) => {
+      try {
+        const exact = await client.listRecords<{
+          sys_id: SnReference;
+          name: SnReference;
+          label: SnReference;
+        }>('sys_db_object', `name=${name}`, ['sys_id', 'name', 'label'], 1);
+
+        if (exact.length === 1) {
+          const t = exact[0];
+          return textResult(
+            [
+              `Table resolved.`,
+              ``,
+              `Name:   ${resolveDisplay(t.name)}`,
+              `Label:  ${resolveDisplay(t.label)}`,
+              `sys_id: ${resolveValue(t.sys_id)}`,
+            ].join('\n'),
+          );
+        }
+
+        const results = await client.listRecords<{
+          sys_id: SnReference;
+          name: SnReference;
+          label: SnReference;
+        }>(
+          'sys_db_object',
+          `nameLIKE${name}^ORlabelLIKE${name}`,
+          ['sys_id', 'name', 'label'],
+          20,
+        );
+
+        if (results.length === 0) {
+          return textResult(`No table found matching "${name}".`);
+        }
+
+        const lines = [
+          `Found ${results.length} table(s) matching "${name}":`,
+          '',
+          ...results.map((t) =>
+            [
+              `• ${resolveDisplay(t.label)} (${resolveDisplay(t.name)})`,
+              `  sys_id: ${resolveValue(t.sys_id)}`,
+            ].join('\n'),
+          ),
+          '',
+          'Tip: re-call with the exact internal name for a single match.',
+        ];
+
+        return textResult(lines.join('\n'));
       } catch (err) {
         return handleError(err);
       }
